Added code for converting roman to integer numbers.
Basic idea is to check if the previous character is less than
the future next character. If so minus it from total
else add it to the total. Make sure to check for boundary condition
when i = length of the given roman number. Since there are no values
there are chances of array index out of bounds.